Alternative SMS Design
This is the SMS design I came up with when I attempted to make my own
minimal theme a month or so ago. It's a bit more square but a lot more
simple in design. I've made it a seperate theme folder so that people
can have a look and see if they prefer it or want to adapt it in some
way.